Replace the localesBarrier with a call to chpl_comm_barrier()
The localesBarrier was a hacky and not scalable barrier. It's hobbled
because we can't use many language features since we're so early in
initialization. The original commit (340ce913d6) noted that it'd be nice
to use `chpl_comm_barrier()`. That wasn't possible at the time because
the comm barrier was already in-use, but that was changed in 8533 and
now we can just use chpl_comm_barrier. This removes some crufty code,
and should be a lot more scalable, though it's only called once so we
probably won't notice.

__EDIT__:

Turns out this doesn't work because the locales array is set up during
runtime init and non-zero locales are still waiting on a setup barrier.
We could probably do some refactoring here, but this does not work as-is