Improve task placement for local bounded coforalls
Apply our patch to help implement a simple task placement/resetting
policy to improve task affinity. This applies 12868 to qthreads 1.16

From the original commit:
""""
This implements a simple task placement/resetting policy to improve task
affinity.

Task placement is currently determined by qthreads. The nemesis
scheduler (our default) does a round-robin assignment of tasks to
workers, where each worker maintains a current index to spawn to (and
there's a global counter for tasks spawned from a non-qthread.) This
task placement results in good affinity for consecutive loops that use
all the cores so long as there's no interleaving tasks, but it gets
affinity wrong when `tasks < cores` or if there's interleaving tasks
created. For a stream variant that intentionally creates tasks between
first-touch and the kernel, this results in a 2.5x performance hit.

Here we just reset the worker spawn index before spawning tasks for a
local bounded coforall when spawning at least dataParTasksPerLocale
tasks. This gives us good affinity for consecutive forall loops that
don't use all the cores, and ensures that all forall loops get the same
affinity even if there's interleaving tasks. For example:

```chpl
forall a in A do a += 1;
cobegin { something(); something();}
forall a in A do a += 1;
```

will now result in the forall loops having the same task affinity.

This improves the performance of stream with interleaving tasks.

| interleaving tasks | before perf | current perf |
| ------------------ | ----------- | ------------ |
| none               |  84 GB/s    |  84 GB/s     |
| numCores/2         |  32 GB/s    |  84 GB/s     |
| random             | ~45 GB/s    |  84 GB/s     |

This also improves prk-stencil by ~5% (and will also allow us to remove
some hacks used to prevent interleaving tasks from hurting performance.)

I view this as as a stop-gap solution and I probably won't contribute
this change upstream. I think the real solution is to have a bulk/vector
task spawn API for local tasks and take care of affinity there using the
existing qthreads API to spawn to to a specific worker -- 9614.
"""

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>