Update 0 size chpl_good_alloc_size under jemalloc
Previously, chpl_good_alloc_size would return 0 when the size was 0
under jemalloc. However, if somebody is calling good_alloc_size they're
about to allocate memory anyways and 0-sized allocations will end up
being 8 bytes, so just return that. It doesn't really make sense for
code to be doing `chpl_good_alloc_size(0)`, but if it is return a more
accurate answer. This matches the behavior of the system allocator on
osx.