Optimize Block dist scan
Use local DR arrays for the inputReady array and the array that stores
the locale-private totals. We want these to live on the initiating
locale to increase parallelism. Previously, the initiating locale would
do a blocking on + sync to see if the input was ready, and then a
blocking GET of the locale private total for each locale. Now the remote
locales will will PUT the locale private total and update input ready.
This reduces the amount of work and communication required by the
iniating locale.

Also have the initiating locale update the remote outputReady elements
in parallel by using a coforall+on instead of operating serially.

Lasly use BlockDist instead of ReplicatedDist for the outputReady array.
This avoids replicated domain and array creation.

TODO:
 - add test for concurrent scans
 - add test for scans over a subset of the locales
 - add test when initiating locale not in target locales (and optimize)