BACKPORT/2021.3.0: support Cray PMI independent of aries
This commit is a backport to GASNet-2021.3.0 of the one described
below, with one or more prior commits squashed in to satisfy
dependencies.  This is NOT the canonical version of this work.

--

This commit removes the probe for Cray PMI support from the probe
for aries-conduit, instead integrating it into the (revised) logic
used to detect the three other flavors of PMI implementation, as used
by other conduits via the generic spawner infrastructure.

This is a necessary step toward supporting efficient job launch for
ofi and/or ucx conduits on current Cray EX series systems with the
Slingshot-10 network.  Consequently, the new version of the Cray PMI
probe deals with the use of `libpmi2` on the EX series, where the XC
series (and gemini-conduit predecessors) used `libpmi`.

Tested extensively on Theta@ALCF (Cray XC / Aries) and Spock@OLCF
(Cray EX / Slingshot-10).  This includes builds on both systems
with the `cc` compiler wrapper as well as with "bare" compilers.