Create a git patch for .good file mismatches
Create a diff that can be `git apply`d when the testing infra has a diff
failure. This is a prototype to accept .good file changes with `git
apply`. It's pretty straight forward, the only slightly off part is that
we have to use `--no-index` in order to get a diff with something no in
source control so we have to build the index paths back up ourselves.

This doesn't work with paratest, and I'm not sure what the easiest way
is to get it working. Maybe instead of creating the "apply log" on the
fly, we could keep track of which files failed to diff and then create
it afterwards. This would also be interesting if there was enough output
in the test log to create the patch afterwards..

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>